re-work summary page flow
