+++ src/server/routes/create/resources/detail/page.njk
@@ -60,7 +60,7 @@
           {% endfor %}
 
           {{ appButton({
-            text: 'Submit request',
+            text: 'Next',
             loader: {
               name: "request-button"
             }
