+++ src/server/routes/create/resources/detail/sqs_queues/{uuid~}/route.js
@@ -5,17 +5,17 @@ import { sortByName } from '#server/common/helpers/sort/sort-by-name.js'
 import formEngine from '#server/plugins/form-engine/form-engine.js'
 import { repositoryNameValidation } from '@defra/cdp-validation-kit'
 import Joi from 'joi'
-import handleNoBasket from '../../ext/handleNoBasket.js'
-import provideLayoutContext from '../../ext/provideLayoutContext.js'
-import createEnvironmentOptions from '../../domain/create-environment-options.js'
-import deduplicationScopeOptions from '../../domain/deduplication-scope-options.js'
-import fifoThroughputLimitOptions from '../../domain/fifo-throughput-limit-options.js'
+import handleNoBasket from '../../../ext/handleNoBasket.js'
+import provideLayoutContext from '../../../ext/provideLayoutContext.js'
+import createEnvironmentOptions from '../../../domain/create-environment-options.js'
+import deduplicationScopeOptions from '../../../domain/deduplication-scope-options.js'
+import fifoThroughputLimitOptions from '../../../domain/fifo-throughput-limit-options.js'
 import {
   Resources,
   getBasketResource,
   updateBasketResource
-} from '../../domain/basket.js'
-import { options as parentOptions } from '../../route.js'
+} from '../../../domain/basket.js'
+import * as parent from '../../route.js'
 
 export function register(routePath) {
   return [
@@ -25,7 +25,7 @@ export function register(routePath) {
         route: {
           path: routePath,
           options: {
-            auth: parentOptions.auth
+            auth: parent.options.auth
           }
         },
