+++ src/server/routes/create/resources/domain/create-environment-options.js
